Share method functionality
Moved functions to handle method panels on the Family and Group forms to one shared place.
