+++ src/clincoded/static/components/family_curation.js
@@ -8,6 +8,7 @@ var form = require('../libs/bootstrap/form');
 var globals = require('./globals');
 var curator = require('./curator');
 var RestMixin = require('./rest').RestMixin;
+var methods = require('./methods');
 
 var CurationMixin = curator.CurationMixin;
 var RecordHeader = curator.RecordHeader;
@@ -37,11 +38,6 @@ var FamilyCuration = React.createClass({
 
     // Keeps track of values from the query string
     queryValues: {},
-
-    instance: {
-        // If we're adding a family to a group, keep a flattened group here so we can update it when the family is submitted
-        group: {},
-    },
 
     getInitialState: function() {
         return {
@@ -171,7 +167,7 @@ var FamilyCuration = React.createClass({
         // If a method and/or segregation object was created (at least one method/segregation field set), assign it to the family.
         // If writing multiple family objects, reuse the one we made, but assign new methods and segregations because each family
         // needs unique objects here.
-        var newMethod = this.createMethod();
+        var newMethod = methods.create.call(this);
         if (newMethod) {
             writerFamily.method = newMethod;
         }
@@ -479,56 +475,6 @@ var FamilyCuration = React.createClass({
         }
     },
 
-    // Create method object based on the form values
-    createMethod: function() {
-        var newMethod = {};
-        var value1, value2;
-
-        // Put together a new 'method' object
-        value1 = this.getFormValue('prevtesting');
-        if (value1 !== 'none') {
-            newMethod.previousTesting = value1 === 'Yes';
-        }
-        value1 = this.getFormValue('prevtestingdesc');
-        if (value1) {
-            newMethod.previousTestingDescription = value1;
-        }
-        value1 = this.getFormValue('genomewide');
-        if (value1 !== 'none') {
-            newMethod.genomeWideStudy = value1 === 'Yes';
-        }
-        value1 = this.getFormValue('genotypingmethod1');
-        value2 = this.getFormValue('genotypingmethod2');
-        if (value1 !== 'none' || value2 !== 'none') {
-            newMethod.genotypingMethods = _([value1, value2]).filter(function(val) {
-                return val !== 'none';
-            });
-        }
-        value1 = this.getFormValue('entiregene');
-        if (value1 !== 'none') {
-            newMethod.entireGeneSequenced = value1 === 'Yes';
-        }
-        value1 = this.getFormValue('copyassessed');
-        if (value1 !== 'none') {
-            newMethod.copyNumberAssessed = value1 === 'Yes';
-        }
-        value1 = this.getFormValue('mutationsgenotyped');
-        if (value1 !== 'none') {
-            newMethod.specificMutationsGenotyped = value1 === 'Yes';
-        }
-        value1 = this.getFormValue('specificmutation');
-        if (value1) {
-            newMethod.specificMutationsGenotypedMethod = value1;
-        }
-        value1 = this.getFormValue('additionalinfomethod');
-        if (value1) {
-            newMethod.additionalInformation = value1;
-        }
-        newMethod.dateTime = moment().format();
-
-        return Object.keys(newMethod).length ? newMethod : null;
-    },
-
     // Create segregation object based on the form values
     createSegregation: function(newFamily, variants) {
         var newSegregation = {};
@@ -689,6 +635,8 @@ var FamilyCuration = React.createClass({
     render: function() {
         var annotation = this.state.annotation;
         var gdm = this.state.gdm;
+        var family = this.state.family;
+        var method = (family.method && Object.keys(family.method).length) ? family.method : {};
         var submitErrClass = 'submit-err pull-right' + (this.anyFormErrors() ? '' : ' hidden');
 
         // Get the query strings. Have to do this now so we know whether to render the form or not. The form
@@ -729,7 +677,7 @@ var FamilyCuration = React.createClass({
                                         </PanelGroup>
                                         <PanelGroup accordion>
                                             <Panel title="Family — Methods" open>
-                                                {FamilyMethods.call(this)}
+                                                {methods.render.call(this, method)}
                                             </Panel>
                                         </PanelGroup>
                                         <PanelGroup accordion>
@@ -956,83 +904,6 @@ var FamilyDemographics = function() {
                     <option>Years</option>
                 </Input>
             </div>
-        </div>
-    );
-};
-
-
-// Methods family curation panel. Call with .call(this) to run in the same context
-// as the calling component.
-var FamilyMethods = function() {
-    var family = this.state.family;
-    var method = (family.method && Object.keys(family.method).length) ? family.method : {};
-
-    return (
-        <div className="row">
-            <Input type="select" ref="prevtesting" label="Previous Testing:" defaultValue="none" value={curator.booleanToDropdown(method.previousTesting)}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Yes</option>
-                <option>No</option>
-            </Input>
-            <Input type="textarea" ref="prevtestingdesc" label="Description of Previous Testing:" rows="5" value={method.previousTestingDescription}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group" />
-            <Input type="select" ref="genomewide" label="Genome-wide Study?:" defaultValue="none" value={curator.booleanToDropdown(method.genomeWideStudy)}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Yes</option>
-                <option>No</option>
-            </Input>
-            <h4 className="col-sm-7 col-sm-offset-5">Genotyping Method</h4>
-            <Input type="select" ref="genotypingmethod1" label="Method 1:" handleChange={this.handleChange} defaultValue="none" value={method.genotypingMethods && method.genotypingMethods[0] ? method.genotypingMethods[0] : null}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Exome sequencing</option>
-                <option>Genotyping</option>
-                <option>HRM</option>
-                <option>PCR</option>
-                <option>Sanger</option>
-                <option>Whole genome shotgun sequencing</option>
-            </Input>
-            <Input type="select" ref="genotypingmethod2" label="Method 2:" defaultValue="none" value={method.genotypingMethods && method.genotypingMethods[1] ? method.genotypingMethods[1] : null}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group" inputDisabled={this.state.genotyping2Disabled}>
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Exome sequencing</option>
-                <option>Genotyping</option>
-                <option>HRM</option>
-                <option>PCR</option>
-                <option>Sanger</option>
-                <option>Whole genome shotgun sequencing</option>
-            </Input>
-            <Input type="select" ref="entiregene" label="Entire gene sequenced?:" defaultValue="none" value={curator.booleanToDropdown(method.entireGeneSequenced)}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Yes</option>
-                <option>No</option>
-            </Input>
-            <Input type="select" ref="copyassessed" label="Copy number assessed?:" defaultValue="none" value={curator.booleanToDropdown(method.copyNumberAssessed)}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Yes</option>
-                <option>No</option>
-            </Input>
-            <Input type="select" ref="mutationsgenotyped" label="Specific Mutations Genotyped?:" defaultValue="none" value={curator.booleanToDropdown(method.specificMutationsGenotyped)}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group">
-                <option value="none">No Selection</option>
-                <option disabled="disabled"></option>
-                <option>Yes</option>
-                <option>No</option>
-            </Input>
-            <Input type="textarea" ref="specificmutation" label="Method by which Specific Mutations Genotyped:" rows="5" value={method.specificMutationsGenotypedMethod}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group" />
-            <Input type="textarea" ref="additionalinfomethod" label="Additional Information about Family Method:" rows="8" value={method.additionalInformation}
-                labelClassName="col-sm-5 control-label" wrapperClassName="col-sm-7" groupClassName="form-group" />
         </div>
     );
 };
